pyproject.toml
[project]
name = "ejemplo-crewai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "openai>=1.97.0",
    "streamlit-js-eval>=0.1.7",
    "streamlit>=1.47.0",
    "langchain-openai>=0.3.28",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.4",
]
